requirements.txt
pytest
requests
PyYAML
jsonschema
pytest-html